wx_vc9.sln opened via Visual Studio 2012 and converted --> UpgradeLog.XML
The $(TargetName) and the $(TargetPath) reported as not matching
the Librarian's OutputFile property.

Compliation started with the Debug and Release configurations.
For each project and for its Debug, the Configuration Properties -
General - Target Name changed from "$(ProjectName)" to "$(ProjectName)d".
This way, many warnings disappeared and many projects compiled without
warnings.